+++ src/main/kotlin/com/dvd/intellij/d2/ide/completion/D2ShapesTypesCompletionContributor.kt
@@ -1,10 +1,14 @@
 package com.dvd.intellij.d2.ide.completion
 
 import com.dvd.intellij.d2.ide.utils.Shapes
-import com.dvd.intellij.d2.ide.utils.TruePattern
+import com.dvd.intellij.d2.lang.D2ElementTypes
+import com.dvd.intellij.d2.lang.psi.impl.D2ShapeDefinitionImpl
 import com.intellij.codeInsight.completion.*
 import com.intellij.codeInsight.lookup.LookupElementBuilder
 import com.intellij.icons.AllIcons
+import com.intellij.patterns.PlatformPatterns
+import com.intellij.psi.util.childrenOfType
+import com.intellij.psi.util.parentOfType
 import com.intellij.util.ProcessingContext
 
 class D2ShapesTypesCompletionContributor : CompletionContributor() {
@@ -15,13 +19,22 @@ class D2ShapesTypesCompletionContributor : CompletionContributor() {
   }
 
   init {
-    extend(CompletionType.BASIC, TruePattern, object : CompletionProvider<CompletionParameters>() {
-      override fun addCompletions(
-        parameters: CompletionParameters,
-        context: ProcessingContext,
-        result: CompletionResultSet
-      ) = result.addAllElements(shapes)
-    })
+    extend(
+      CompletionType.BASIC,
+      PlatformPatterns.psiElement(D2ElementTypes.IDENTIFIER),
+      object : CompletionProvider<CompletionParameters>() {
+        override fun addCompletions(
+          parameters: CompletionParameters,
+          context: ProcessingContext,
+          result: CompletionResultSet
+        ) {
+          val parent = parameters.position.parentOfType<D2ShapeDefinitionImpl>() ?: return
+          parent.childrenOfType<D2ShapeDefinitionImpl>().firstOrNull { it.text == "shape" } ?: return
+
+          result.addAllElements(shapes)
+        }
+      }
+    )
   }
 
 }