+++ src/main/kotlin/com/dvd/intellij/d2/ide/format/D2FormatterService.kt
@@ -42,6 +42,7 @@ class D2FormatterService : AsyncDocumentFormattingService() {
         }
       }
 
+      override fun isRunUnderProgress(): Boolean = true
       override fun cancel(): Boolean = true
     }
   }